+++ lib/public_activity/config.rb
@@ -6,14 +6,6 @@ module PublicActivity
     include ::Singleton
     attr_accessor :enabled, :table_name
 
-    @@orm = :active_record
-
-    def initialize
-      # Indicates whether PublicActivity is enabled globally
-      @enabled    = true
-      @table_name = "activities"
-    end
-
     # Evaluates given block to provide DSL configuration.
     # @example Initializer for Rails
     #   PublicActivity::Config.set do
@@ -24,14 +16,15 @@ module PublicActivity
     def self.set &block
       b = Block.new
       b.instance_eval &block
-      @@orm = b.orm unless b.orm.nil?
+      orm(b.orm) unless b.orm.nil?
       instance
-      instance.instance_variable_set(:@enabled,    b.enabled)     unless  b.enabled.nil?
+      Thread.current[:public_activity_enabled] = b.enabled unless b.enabled.nil?
+      Thread.current[:public_activity_table_name] = b.table_name unless b.enabled.nil?
     end
 
     # Set the ORM for use by PublicActivity.
     def self.orm(orm = nil)
-      @@orm = (orm ? orm.to_sym : false) || @@orm
+      Thread.current[:public_activity_orm] = (orm ? orm.to_sym : false) || Thread.current[:public_activity_orm]
     end
 
     # alias for {#orm}
@@ -44,6 +37,10 @@ module PublicActivity
     # @see Config#orm
     def orm(orm=nil)
       self.class.orm(orm)
+    end
+
+    def table_name
+      Thread.current[:public_activity_orm] || "activities"
     end
 
     # Provides simple DSL for the config block.
@@ -63,7 +60,7 @@ module PublicActivity
       # Sets the table_name
       # for the model
       def table_name(name = nil)
-        PublicActivity.config.table_name = name
+        @table_name = (name.nil? ? @table_name : name)
       end
     end
   end